refactor: combine viewer
