+++ packages/skeleton-helper/src/SkeletonViewer.ts
@@ -1,49 +1,99 @@
 import {
+  Color,
   Engine,
   Entity,
+  Material,
   Matrix,
   MeshRenderer,
   ModelMesh,
   PrimitiveMesh,
   Quaternion,
+  RenderQueueType,
+  Script,
+  Shader,
   SkinnedMeshRenderer,
   Vector3
 } from "oasis-engine";
-import { SkeletonHelper } from "./SkeletonManager";
 
-export class SkeletonViewer {
-  engine: Engine;
-  debugMesh: MeshRenderer[] = [];
-  skin: SkinnedMeshRenderer;
-  manager: SkeletonHelper;
+Shader.create(
+  "skeleton-viewer",
+  `
+  attribute vec3 POSITION;
+  attribute vec3 NORMAL;
 
-  constructor(engine: Engine, skin: SkinnedMeshRenderer, manager: SkeletonHelper) {
-    this.engine = engine;
-    this.skin = skin;
-    this.manager = manager;
+  uniform mat4 u_MVPMat;
+  uniform mat4 u_normalMat;
 
-    if (!this.skin.jointNodes) {
-      this.skin.update(0);
+  varying vec3 v_normal;
+
+  void main(){
+      gl_Position = u_MVPMat * vec4( POSITION , 1.0 );;
+      v_normal = normalize( mat3(u_normalMat) * NORMAL );
+  }`,
+  `
+      uniform vec3 u_colorMin;
+      uniform vec3 u_colorMax;
+      varying vec3 v_normal;
+
+      void main(){
+        float ndl = dot(v_normal, vec3(0, 1, 0)) * 0.5 + 0.5;
+        vec3 diffuse = mix(u_colorMin, u_colorMax, ndl);
+        gl_FragColor = vec4(diffuse, 1.0);
+      }
+      `
+);
+const materialMap = new Map<Engine, Material>();
+
+export class SkeletonViewer extends Script {
+  material: Material;
+
+  // Config
+  midStep: number = 0.2;
+  midWidthScale: number = 0.1;
+  ballSize: number = 0.25;
+  scaleFactor: number = 0.85;
+  colorMin: Color = new Color(0.35, 0.35, 0.35, 1);
+  colorMax: Color = new Color(0.7, 0.7, 0.7, 1);
+
+  private _debugMesh: MeshRenderer[] = [];
+
+  constructor(entity: Entity) {
+    super(entity);
+
+    const engine = entity.engine;
+    if (!materialMap.get(engine)) {
+      const material = new Material(entity.engine, Shader.find("skeleton-viewer"));
+      material.renderState.rasterState.depthBias = -100000000;
+      material.renderQueueType = RenderQueueType.Transparent;
+      materialMap.set(engine, material);
+    }
+
+    this.material = materialMap.get(engine);
+    this.material.shaderData.setColor("u_colorMin", this.colorMin);
+    this.material.shaderData.setColor("u_colorMax", this.colorMax);
+
+    const skinnedMeshRenderers = [];
+    this.entity.getComponentsIncludeChildren(SkinnedMeshRenderer, skinnedMeshRenderers);
+    for (let i = 0; i < skinnedMeshRenderers.length; i++) {
+      const renderer = skinnedMeshRenderers[i];
+      if (renderer.skin) {
+        this._showSkeleton(renderer);
+      }
     }
   }
 
-  update(): void {
-    this.destroy();
-    this._showSkeleton();
-  }
-
-  destroy(): void {
-    for (let i = 0, length = this.debugMesh.length; i < length; i++) {
-      this.debugMesh[i].destroy();
+  onDestroy(): void {
+    for (let i = 0, length = this._debugMesh.length; i < length; i++) {
+      this._debugMesh[i].destroy();
     }
-    this.debugMesh.length = 0;
+    this._debugMesh.length = 0;
   }
 
   private _createSpur(direction: Vector3): ModelMesh {
     const mesh = new ModelMesh(this.engine);
     const length = direction.length();
-    const midLength = this.manager.midStep * length;
-    const midHalfWidth = length * this.manager.midWidthScale;
+    const midLength = this.midStep * length;
+    const midHalfWidth = length * this.midWidthScale;
 
     const positions: Vector3[] = new Array(24);
     const normals: Vector3[] = new Array(24);
@@ -116,8 +166,12 @@ export class SkeletonViewer {
     return mesh;
   }
 
-  private _showSkeleton(): void {
-    const joints = this.skin.jointNodes;
+  private _showSkeleton(renderer: SkinnedMeshRenderer): void {
+    if (!renderer.jointNodes) {
+      renderer.update(0);
+    }
+
+    const joints = renderer.jointNodes;
 
     const spheres: Entity[][] = [];
 
@@ -130,13 +184,13 @@ export class SkeletonViewer {
       // 球
       const entity = joint.createChild();
       const renderer = entity.addComponent(MeshRenderer);
-      renderer.mesh = PrimitiveMesh.createSphere(this.engine, this.manager.ballSize, 16);
-      renderer.setMaterial(this.manager.material);
+      renderer.mesh = PrimitiveMesh.createSphere(this.engine, this.ballSize, 16);
+      renderer.setMaterial(this.material);
       renderer.priority = 1;
 
       spheres.push([entity, joint]);
 
-      this.debugMesh.push(renderer);
+      this._debugMesh.push(renderer);
 
       // 连接体
       for (let j = 0; j < joint.childCount; j++) {
@@ -152,11 +206,11 @@ export class SkeletonViewer {
 
         const entity = joint;
         const renderer = entity.addComponent(MeshRenderer);
-        renderer.setMaterial(this.manager.material);
+        renderer.setMaterial(this.material);
         renderer.mesh = this._createSpur(direction);
         renderer.priority = 1;
 
-        this.debugMesh.push(renderer);
+        this._debugMesh.push(renderer);
       }
     }
 
@@ -173,7 +227,7 @@ export class SkeletonViewer {
         base = base.parent;
       }
 
-      const scale = 0.5 * maxLength * Math.pow(this.manager.scaleFactor, count);
+      const scale = 0.5 * maxLength * Math.pow(this.scaleFactor, count);
       const worldScale = sphere.transform.lossyWorldScale;
       sphere.transform.setScale(scale / worldScale.x, scale / worldScale.y, scale / worldScale.z);
     }
