+++ packages/skeleton-helper/src/index.ts
@@ -1,1 +1,1 @@
-export { SkeletonHelper } from "./SkeletonManager";
+export { SkeletonViewer } from "./SkeletonViewer";
